Refactor: Isolate Prisma client to server-only
