+++ packages/web/package.json
@@ -60,6 +60,7 @@
     "postcss": "^8.4.32",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "server-only": "^0.0.1",
     "stripe": "^20.0.0",
     "tailwind-merge": "^2.2.0",
     "tailwindcss": "^3.4.0",
